radify_api/requirements.txt
# FastAPI framework and server
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Database and ORM
sqlalchemy==2.0.29
mysql-connector-python==8.4.0

# Pydantic and password handling
pydantic==2.6.4
passlib[bcrypt]==1.7.4

# JWT Authentication
python-jose[cryptography]==3.3.0

# Form and email handling
python-multipart==0.0.9
email-validator==2.1.1

# CrewAI and LangChain dependencies
crewai
# langchain
# langchain-core
# langchain-community
openai
crewai-tools
python-dotenv

# Google OAuth and Drive API
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.125.0
